Introduced 'provenance serve' command.
