+++ executables/provenance
@@ -82,6 +82,9 @@ approve = subparsers.add_parser('approve',
 approve.add_argument('filepath', 
     help='Where the file resides.')
 
+serve = subparsers.add_parser('serve',
+    help='Start a webserver that allows browsing through provenance.')
+
 args = parser.parse_args()
 if args.command == 'discover':
     niprov.discover(args.root)
@@ -107,3 +110,5 @@ elif args.command == 'todo':
     niprov.markedForApproval()
 elif args.command == 'approve':
     niprov.approve(args.filepath)
+elif args.command == 'serve':
+    niprov.serve()
